test_sampler_uniforms
